Commit many new changes to osm2pgsql, including: - list of tags read from file - slim mode works again - relations properly supported - more efficient DB usage
It includes some restructuring of the code, in particular, the output module
manages the mid-level now and the main program doesn't call it at all. This
moves many of the previous hacks to the output module which can manage the
mid-level as appropriate for its output (i.e. slim mode requires different
semantics from ram mode)