This commit introduces a new and consistant rendering style for borders based on admin_level. It falls back to border_type if no explicit admin_level is given.
It assumes that nation / country are equivalent to admin_level=2,
region / state are equivalent to admin_level=4, county / district are
equivalent to admin_level=6, municipality / city are equivalent to
admin_level=8 and suburb is equivalent to admin_level=10. This is
consistent with the wiki and the majority of the ways tagged with both
border_type and admin_level.

The new rendering might need minor tweaks, especially for zoom level 12
and low zoom tiles. Never the less it should fix trac tickets #637,
#1150, #1184 and #1332.