Introducing a new /browse/{node,way,relation,changeset}/{id}/map page which is slippy map covering the screen showing the map data that is in the small map at /browse/{node,way,relation,changeset}/{id}.
/browse/{node,way,relation,changeset}/{id} now contains a link to both
"View Larger Map" which has been changed to link to this new map
feature, and "View on main map" which as before links to the area
showing the feature on the main slippy map.

/browse/{node,way,relation,changeset}/{id}/map supports permalinks
using its own hacked version of updateLocation() (which doesn't set a
cookie), but shortlinks are not supported. The Edit/History/Export
links link back to the main map.