This is the initial implementation of login via OpenID
This is the patch presented in trac ticket #2500

With this commit, it is possible to assosciate an openID
to an existing OSM account. Once associated, it is possible
to either login via OpenID, or as always via username and password.

Other aspects, such as account creation and the need for a valid email
are unchanged.