[osm2pgsql] - Create a conditional index on planet_osm_polygon where building is null
Over time, the planet_osm_polygon table has increased in size considerably to currently 17Gb for the full import.
During low zoom rendering, a good portion of time is spent with polygon querries. 
Roughly 80% of these polygons are buildings.

Therefor adding an index on the geometry column for when building is null, can help speed up low zoom rendering by quite a bit.
However, to fully take advantage of this index, the style sheets need to take it into account.